Moved types from old shortUrlsListParams reducer, to the data index file
